Factor out usePermalinkMember
